add lib project, tests project and solution
