+++ FastMatrix.Net.Test/Usings.cs
@@ -1,0 +1,2 @@
+global using Xunit;
+global using FastMatrix.Net;
