+++ FastMatrix.Net.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FastMatrix.Net", "FastMatrix.Net\FastMatrix.Net.csproj", "{AA6BC434-7B3D-4E32-B265-F22DDBB4798B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FastMatrix.Net.Test", "FastMatrix.Net.Test\FastMatrix.Net.Test.csproj", "{66FD708A-6720-4AFA-9387-6C18EB2E6832}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{AA6BC434-7B3D-4E32-B265-F22DDBB4798B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AA6BC434-7B3D-4E32-B265-F22DDBB4798B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AA6BC434-7B3D-4E32-B265-F22DDBB4798B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AA6BC434-7B3D-4E32-B265-F22DDBB4798B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{66FD708A-6720-4AFA-9387-6C18EB2E6832}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{66FD708A-6720-4AFA-9387-6C18EB2E6832}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{66FD708A-6720-4AFA-9387-6C18EB2E6832}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{66FD708A-6720-4AFA-9387-6C18EB2E6832}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
